dev-service/requirements.txt
flask==3.0.0
PyGithub==2.1.1
python-dotenv==1.0.0
google-genai>=1.0.0
openai==1.12.0
gunicorn==21.2.0
